get rid of pthread mutex
