+++ packages/bun-usockets/src/loop.c
@@ -42,7 +42,7 @@ void us_internal_loop_data_init(struct us_loop_t *loop, void (*wakeup_cb)(struct
 
     loop->data.closed_connecting_head = 0;
     loop->data.dns_ready_head = 0;
-    pthread_mutex_init(&loop->data.mutex, 0);
+    loop->data.mutex = 0;
 
     loop->data.wakeup_async = us_internal_create_async(loop, 1, 0);
     us_internal_async_set(loop->data.wakeup_async, (void (*)(struct us_internal_async *)) wakeup_cb);
@@ -170,10 +170,10 @@ void us_internal_handle_low_priority_sockets(struct us_loop_t *loop) {
 }
 
 void us_internal_handle_dns_results(struct us_loop_t *loop) {
-    pthread_mutex_lock(&loop->data.mutex);
+    Bun__lock(&loop->data.mutex);
     struct us_connecting_socket_t *s = loop->data.dns_ready_head;
     loop->data.dns_ready_head = NULL;
-    pthread_mutex_unlock(&loop->data.mutex);
+    Bun__unlock(&loop->data.mutex);
 
     while (s) {
         struct us_connecting_socket_t *next = s->next;
